\name{RUnicode-package}
\alias{RUnicode-package}
\alias{RUnicode}
\docType{package}
\title{
  A Function for Unicode Normalization
}
\description{
  normalize unicode characters
}
\details{
\tabular{ll}{
Package: \tab RUnicode\cr
Version: \tab 0.1.0\cr
Date: \tab 2011-04-16\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Arabiki
}
\seealso{
ICU - International Components for Unicode http://site.icu-project.org/
}
