\name{RUnicode-package}
\alias{RUnicode-package}
\alias{RUnicode}
\docType{package}
\title{
Normalize Unicode characters
}
\description{
Normalize Unicode characters of a character vector or factor using ICU library
}
\details{
\tabular{ll}{
Package: \tab RUnicode\cr
Version: \tab 0.1.1\cr
Date: \tab 2013-05-29\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Arabiki
}
\seealso{
ICU - International Components for Unicode http://site.icu-project.org/
}
